+++ src/ts/tools/index.ts
@@ -1,3 +1,4 @@
 export * as analyse from './analyse';
 export * as component from './component';
 export * as swing from './swing';
+export * as parity from './parity';