+++ src/ts/tools/notes/index.ts
@@ -9,6 +9,7 @@ export { default as hitboxStair } from './hitboxStair';
 export { default as hitboxReverseStair } from './hitboxReverseStair';
 export { default as hitboxInline } from './hitboxInline';
 export { default as improperWindow } from './improperWindow';
+export { default as parity } from './parity';
 export { default as doubleDirectional } from './doubleDirectional';
 export { default as visionBlock } from './visionBlock';
 export { default as stackedNote } from './stackedNote';