+++ src/ts/ui/select.ts
@@ -1,11 +1,11 @@
 export const create = (
     id: string,
     label: string,
-    container: string,
+    containerTag: string,
     containerClass: string,
     ...options: { text: string; value: string }[]
 ): HTMLElement => {
-    const htmlContainer = document.createElement(container);
+    const htmlContainer = document.createElement(containerTag);
     htmlContainer.className = containerClass;
 
     const htmlSelect = document.createElement('select');