+++ CHANGELOG.md
@@ -1,10 +1,28 @@
 # Changelog
 
+## v2.1.0 [04-08-2021]
+
+### Added
+
+-   Parity check (EXPERIMENTAL: please send feedback)
+-   2-wide center wall recovery input
+
+### Changed
+
+-   Bomb stack now checks by NJS instead of time
+-   Minor code clean up
+
+### Fixed
+
+-   Varying speed slider not picking up stack + slider
+-   Mistaken 2-wide center wall for negative wall
+-   BeatSaver link not having correct URL
+
 ## v2.0.3 [27-08-2021]
 
 ### Fixed
 
--   BeatSaver download not having correct API link
+-   BeatSaver download not having correct request URL
 
 ## v2.0.2 [08-08-2021]